Move export logic out of CLI module into conda_workspaces.export
Rename ``conda_workspaces/env_export.py`` to
``conda_workspaces/export.py`` and absorb the non-CLI portion of
``cli/workspace/export.py`` into it.  The new module owns both
sides of export:

* Provider — ``multiplatform_export`` / ``_envs_to_dict``, the
  ``conda-workspaces-lock-v1`` exporter plugin callable, unchanged
  from the old ``env_export.py``.
* Consumer — ``resolve_exporter`` / ``run_exporter`` over
  ``conda_context.plugin_manager``, and the three ``Environment``
  builders ``build_from_declared`` / ``build_from_prefix`` /
  ``build_from_lockfile``, one per ``conda workspace export``
  source.

``cli/workspace/export.py`` drops from 331 to 142 lines and now
only inspects ``args``, routes to the builders, and writes output
to stdout / JSON / disk — matching the thin-shim pattern the other
CLI modules follow (``cli/workspace/lock.py`` is 78 lines).

The old ``env_export.py`` name was misleading for the consumer-side
code we were about to colocate: it sounded provider-only ("we are
an exporter plugin") when the module now also owns the consumer
path ("we drive conda's exporter registry").  ``export.py`` covers
both responsibilities the same way ``lockfile.py`` covers both
reader (``CondaLockLoader``) and writer (``generate_lockfile`` /
``merge_lockfiles``) for the lockfile format.

Plugin registration, lockfile docstrings, env_spec docstring,
AGENTS.md, docs, and tests are updated to the new path.  Provider-
side tests live at ``tests/test_export.py`` (renamed) and CLI tests
at ``tests/cli/workspace/test_export.py``.
