+++ conda_workspaces/cli/workspace/export.py
@@ -1,32 +1,13 @@
-"""``conda workspace export`` — via the conda environment exporter plugin.
+"""``conda workspace export`` — argparse shim over :mod:`conda_workspaces.export`.
 
-This command is a thin consumer of conda's built-in environment
-exporter registry (``context.plugin_manager``).  It builds one or
-more :class:`conda.models.environment.Environment` objects, then
-delegates serialisation to whichever ``CondaEnvironmentExporter``
-the user selected via ``--format`` or ``--file`` — default
-``environment-yaml`` / ``environment-json`` exporters live in
-:mod:`conda.plugins.environment_exporters`, our own
-``conda-workspaces-lock-v1`` in :mod:`conda_workspaces.plugin`, and
-any third-party exporter plugged into conda is picked up for free.
+Everything export-related — exporter dispatch, ``Environment``
+builders for the three supported sources, and our own
+``conda-workspaces-lock-v1`` serialisation — lives in
+:mod:`conda_workspaces.export`.  This module only knows about the
+CLI surface: it inspects ``args``, picks a source, picks an
+exporter, and routes the result to stdout / JSON / a file.
 
-Each of the three ``Environment`` sources reuses an existing primitive:
-
-* ``--from-prefix`` — the same :meth:`Environment.from_prefix` +
-  :meth:`Environment.extrapolate` pair that
-  :func:`conda.cli.main_export.execute` uses.
-* ``--from-lockfile`` — our own
-  :class:`~conda_workspaces.lockfile.CondaLockLoader` env-spec plugin
-  (``env_for``), identical to the path conda takes when it reads
-  ``--file conda.lock`` through
-  :meth:`Environment.from_cli_with_file_envs`.
-* (default) ``declared`` — the only source without a conda equivalent,
-  exposed as the public :func:`build_from_declared`.  This is the
-  capability that distinguishes ``conda workspace export`` from
-  ``conda export``: emit an ``environment.yml`` / ``conda.lock`` /
-  anything directly from the manifest without installing or solving
-  first, and the hook point for plugin exporters that want to consume
-  declared-but-unsolved workspace environments directly.
+See :mod:`conda_workspaces.export` for the programmatic API.
 """
 
 from __future__ import annotations
@@ -35,40 +16,22 @@ import json
 import sys
 from typing import TYPE_CHECKING
 
-from conda.base.context import context as conda_context
-from conda.exceptions import CondaValueError, EnvironmentExporterNotDetected
-from conda.models.environment import Environment
-from conda.models.environment import EnvironmentConfig as CondaEnvConfig
-from conda.models.match_spec import MatchSpec
+from conda.exceptions import CondaValueError
 from rich.console import Console
 
-from ...exceptions import (
-    EnvironmentNotFoundError,
-    EnvironmentNotInstalledError,
-    LockfileNotFoundError,
-    PlatformError,
+from ...exceptions import EnvironmentNotFoundError
+from ...export import (
+    build_from_declared,
+    build_from_lockfile,
+    build_from_prefix,
+    resolve_exporter,
+    run_exporter,
 )
-from ...lockfile import CondaLockLoader, lockfile_path
-from ...resolver import resolve_environment
 from . import workspace_context_from_args
 
 if TYPE_CHECKING:
     import argparse
     from pathlib import Path
-
-    from conda.plugins.types import CondaEnvironmentExporter
-
-    from ...context import WorkspaceContext
-    from ...models import WorkspaceConfig
-
-
-DEFAULT_FORMAT = "environment-yaml"
-
-# Canonical key the built-in environment-yaml / environment-json
-# exporters use to roundtrip PyPI deps through ``external_packages``.
-# Older conda releases ship the value without exposing a constant, so
-# we hard-code it here to stay compatible with conda >= 26.1.
-_EXTERNAL_PACKAGES_PYPI_KEY = "pip"
 
 
 def execute_export(
@@ -96,52 +59,22 @@ def execute_export(
     requested_platforms: tuple[str, ...] = tuple(
         getattr(args, "export_platforms", None) or ()
     )
+
     if from_lockfile:
-        # ``CondaLockLoader`` is this package's ``EnvironmentSpecBase``
-        # plugin; ``env_for`` is the same entry point
-        # ``Environment.from_cli_with_file_envs`` uses when conda reads
-        # ``--file conda.lock``, so we get identical ``Environment``
-        # objects without rebuilding the reader.
-        path = lockfile_path(ctx)
-        if not path.is_file():
-            raise LockfileNotFoundError(env_name, path)
-        loader = CondaLockLoader(path)
-        available = loader.available_platforms
-        if requested_platforms:
-            unknown = [p for p in requested_platforms if p not in available]
-            if unknown:
-                raise PlatformError(unknown[0], list(available))
-            requested_set = set(requested_platforms)
-            targets = tuple(p for p in available if p in requested_set)
-        else:
-            targets = available
-        if not targets:
-            raise LockfileNotFoundError(env_name, path)
-        try:
-            envs = [loader.env_for(platform=p, name=env_name) for p in targets]
-        except ValueError as exc:
-            raise LockfileNotFoundError(env_name, path) from exc
+        envs = build_from_lockfile(
+            ctx=ctx,
+            env_name=env_name,
+            requested_platforms=requested_platforms,
+        )
     elif from_prefix:
-        # Same primitives ``conda.cli.main_export.execute`` uses:
-        # ``Environment.from_prefix`` for the host platform and
-        # ``Environment.extrapolate`` to project it onto cross-platform
-        # targets.  The extra work here is only the workspace-prefix
-        # lookup + ``EnvironmentNotInstalledError`` guard.
-        if not ctx.env_exists(env_name):
-            raise EnvironmentNotInstalledError(env_name)
-        prefix_env = Environment.from_prefix(
-            prefix=str(ctx.env_prefix(env_name)),
-            name=env_name,
-            platform=ctx.platform,
+        envs = build_from_prefix(
+            ctx=ctx,
+            env_name=env_name,
+            requested_platforms=requested_platforms,
             from_history=bool(getattr(args, "from_history", False)),
             no_builds=bool(getattr(args, "no_builds", False)),
             ignore_channels=bool(getattr(args, "ignore_channels", False)),
-            channels=list(conda_context.channels),
         )
-        if not requested_platforms or requested_platforms == (ctx.platform,):
-            envs = [prefix_env]
-        else:
-            envs = [prefix_env.extrapolate(p) for p in requested_platforms]
     else:
         envs = build_from_declared(
             config=config,
@@ -207,125 +140,3 @@ def execute_export(
         )
 
     return 0
-
-
-def build_from_declared(
-    *,
-    config: WorkspaceConfig,
-    ctx: WorkspaceContext,
-    env_name: str,
-    requested_platforms: tuple[str, ...],
-) -> list[Environment]:
-    """Resolve declared specs from the manifest into one ``Environment`` per platform.
-
-    Produces an :class:`Environment` with ``requested_packages`` only
-    (no solver, no installed packages required) — the novel piece of
-    ``conda workspace export``.  conda itself has no primitive for
-    "build an :class:`Environment` from a manifest without installing
-    it", which is why this lives here rather than in upstream conda.
-
-    Public because it is the distinguishing capability of
-    ``conda workspace export`` and the natural entry point for
-    third-party exporter plugins (or other tooling) that want to turn
-    a ``conda.toml`` manifest into a list of :class:`Environment`
-    objects without going through the CLI.
-    """
-    try:
-        declared = resolve_environment(config, env_name)
-    except PlatformError:
-        # Manifest declares platforms conda doesn't know; fall back to
-        # the host platform so the export still produces something
-        # useful rather than crashing on validation alone.
-        targets: tuple[str, ...] = (ctx.platform,)
-    else:
-        targets = declared.target_platforms(
-            requested=requested_platforms,
-            fallback=ctx.platform,
-        )
-
-    envs: list[Environment] = []
-    for platform in targets:
-        resolved = resolve_environment(config, env_name, platform)
-
-        requested_packages = [
-            MatchSpec(dep.conda_build_form())
-            for dep in resolved.conda_dependencies.values()
-        ]
-
-        external_packages: dict[str, list[str]] = {}
-        pypi_entries = [
-            str(dep).strip()
-            for dep in resolved.pypi_dependencies.values()
-            if not dep.path and not dep.git and not dep.url
-        ]
-        if pypi_entries:
-            external_packages[_EXTERNAL_PACKAGES_PYPI_KEY] = pypi_entries
-
-        envs.append(
-            Environment(
-                name=env_name,
-                platform=platform,
-                config=CondaEnvConfig(
-                    channels=tuple(ch.canonical_name for ch in resolved.channels),
-                ),
-                requested_packages=requested_packages,
-                external_packages=external_packages,
-            )
-        )
-
-    return envs
-
-
-def resolve_exporter(
-    *,
-    format_name: str | None,
-    file_path: Path | None,
-) -> tuple[CondaEnvironmentExporter, str]:
-    """Look up the plugin exporter to use and return ``(exporter, name)``.
-
-    Thin public wrapper over :attr:`conda.base.context.context.plugin_manager`
-    — no existing workspace class owns exporter dispatch, and conda's
-    ``CondaEnvironmentExporter`` is not ours to extend — kept as a
-    free function matching :func:`build_from_declared` / :func:`run_exporter`.
-
-    Precedence matches ``conda export``:
-
-    1. Explicit ``--format`` wins and is looked up by name or alias.
-    2. Otherwise, if ``--file`` is given, detect by filename pattern.
-    3. Otherwise, default to :data:`DEFAULT_FORMAT`.
-    """
-    pm = conda_context.plugin_manager
-
-    if format_name:
-        exporter = pm.get_environment_exporter_by_format(format_name)
-    elif file_path is not None:
-        try:
-            exporter = pm.detect_environment_exporter(str(file_path))
-        except EnvironmentExporterNotDetected:
-            exporter = pm.get_environment_exporter_by_format(DEFAULT_FORMAT)
-    else:
-        exporter = pm.get_environment_exporter_by_format(DEFAULT_FORMAT)
-    return exporter, exporter.name
-
-
-def run_exporter(
-    exporter: CondaEnvironmentExporter,
-    envs: list[Environment],
-) -> str:
-    """Invoke *exporter*, preferring ``multiplatform_export`` when available.
-
-    Companion to :func:`resolve_exporter`: once an exporter has been
-    selected, this normalises the conda plugin interface's two entry
-    points (``multiplatform_export`` for a list of environments,
-    ``export`` for a single one) and always appends a trailing newline
-    so callers can write the result verbatim.
-    """
-    if exporter.multiplatform_export is not None:
-        content = exporter.multiplatform_export(envs)
-    elif exporter.export is not None:
-        content = exporter.export(envs[0])
-    else:
-        raise CondaValueError(
-            f"Exporter '{exporter.name}' has no registered export method."
-        )
-    return content.rstrip() + "\n"
